- Features:   - Added Distribution.Holtsmark as a typed wrapper over Boost.Math’s heavy-tailed Holtsmark distribution (location/scale form) with full PDF/CDF/quantile coverage and entropy support. - Documentation:   - Linked all references to Swift Numerics’ Real protocol directly to the upstream project documentation for easier navigation.   - Added DocC coverage for Distribution/Geometric and refreshed README, Quickstart, and factory guides with geometric examples and alias tables.   - Renamed HighPrecisionConstants to Constants throughout the guides to reflect the streamlined API. - Tests:   - Added regression coverage for Distribution.Geometric (closed-form PMF/CDF, quantiles, moments) and ensured the dynamic factory’s geometric entry matches the typed wrapper and alias set.

+++ Sources/CBoostBridge/impl/distributions/bs_distribution_helpers.hxx
@@ -172,7 +172,7 @@ float bs_geometric_find_lower_bound_on_p_f(float trials,
 }
 
 double bs_geometric_find_lower_bound_on_p_d(double trials, double alpha) {
-    return bs_wrap<float>([&]{ return geometric_distribution<float>::find_lower_bound_on_p(trials, alpha);});
+    return bs_wrap<double>([&]{ return geometric_distribution<double>::find_lower_bound_on_p(trials, alpha);});
 }
 
 long double bs_geometric_find_lower_bound_on_p_l(long double trials, long double alpha) {
